fix(SerialConsole): Set console width based on columns and add auto-sizing option (#396)
affects: @patternfly/react-console

If the number of columns forces the console to be wider than the viewport, then the parent container was not sizing correctly. Added a min-width style that makes the parent container as wide as the console. Also fixed an issue with resizing and added an option to auto-size using the autoFit property.

ISSUES CLOSED: https://github.com/patternfly/patternfly-react/issues/380