Lint md code blocks 3469 (#3474)
* fix(linting): added eslint for *.md files

linting all code blocks in any markdown file. We are only producing warnings so not to break the build. (This will change)

Also fixed all hard errors in this commit so not to break the build.

* fix(linting): added 'dist' to ignore file.

* fix(linting): pr comment changes
Change enclosing <div> tag to empty tag in README file
moved linting to run in circleCI
fixed 'variant' line in README file.

* fix(linting): change eslint rule
make sure that circleCI step fails when there are 'errors'

* fix(linting): moved lint:md to linting steps in circleCI

* fix(linting): back to warnings

* fix(linting): changed name on circleCI md lint step.