Page resize observer & Toolbar/PageHeaderTools responsiveness based on page width (#6827)
- Added a page resize observer to the Page component
- Adds **width**, and **getBreakpoint** to the `PageContext`. The width is the width of the Page (pf-c-page) container. Children components can use the context to base logic around that (like the Toolbar components in this PR). The **getBreakpoint** function defaults to the one in `util.ts`, but can be overwritten if desired.
- Adds a `pf-m-breakpoint-[default|sm|md|lg|xl|2xl]` class as a sibling to the Page (pf-c-page) container. This allows for CSS to target page based breakpoints [as described in this issue](https://github.com/patternfly/patternfly/issues/4622)

A new util function `getBreakpoint` was added to return the breakpoint for a given width.

The `util.ts` **formatBreakpointMods** function was updated to accept a `breakpoint` argument.
Before, you would have a string of class names that are media query based, i.e. `pf-m-hidden pf-m-visible-on-md pf-m-hidden-on-lg`, or `pf-m-spacer-none pf-m-spacer-md-on-md`, etc. If the `breakpoint` argument is supplied, we instead return now a single class based on the known width.